extraction_docker/requirements.txt
attrs==21.2.0
Automat==20.2.0
azure-core==1.20.1
azure-storage-blob==12.9.0
blis==0.4.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
characteristic==14.3.0
charset-normalizer==2.0.7
click==8.0.3
constantly==15.1.0
cryptography==35.0.0
cycler==0.11.0
cymem==2.0.6
deskew==0.10.37
fuzzywuzzy==0.18.0
hyperlink==21.0.0
idna==3.3
imageio==2.10.4
importlib-metadata==4.8.2
imutils==0.5.4
incremental==21.3.0
isodate==0.6.0
Jinja2==3.0.3
kiwisolver==1.3.2
klein==21.8.0
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib==3.4.3
msrest==0.6.21
murmurhash==1.0.6
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
packaging==21.2
pandas==1.3.4
pathy==0.6.1
Pillow==8.4.0
plac==1.1.3
preshed==3.0.6
price-parser==0.3.4
psycopg2-binary==2.9.2
pycparser==2.21
pydantic==1.8.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.26.0
requests-oauthlib==1.3.0
scikit-image==0.18.3
scipy==1.7.2
six==1.16.0
smart-open==5.2.1
srsly==2.4.2
thinc==8.0.13
tifffile==2021.11.2
tqdm==4.62.3
Tubes==0.2.0
Twisted==21.7.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
zipp==3.6.0
zope.interface==5.4.0
azure-keyvault-secrets
azure-identity
rapidfuzz
networkx
regex